user-service/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
pika==1.3.2
